change 6th argument of gfs_digest_calculate_local() from "unsigned *" to "size_t *" for x86_64
git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm/trunk@1302 4a5e1ece-6a47-475c-b646-731d720b41d0

+++ lib/libgfarm/gfarm/gfs_pio_misc.c
@@ -926,7 +926,7 @@ gfs_rename(const char *from_url, const char *to_url)
 
 static char *
 digest_calculate(char *filename,
-		 char **digest_type, char *digest_string, unsigned *md_len_p,
+		 char **digest_type, char *digest_string, size_t *md_len_p,
 		 file_offset_t *filesizep)
 {
 	int fd, i, rv;
@@ -964,7 +964,7 @@ gfs_pio_set_fragment_info_local(char *filename,
 {
 	char *digest_type;
 	char digest_value_string[EVP_MAX_MD_SIZE * 2 + 1];
-	unsigned int digest_len;
+	size_t digest_len;
 	file_offset_t filesize;
 	char *e = NULL;
 	struct gfarm_file_section_info fi;
